sp_externalities

Function set_and_run_with_externalities

pub fn set_and_run_with_externalities<F, R>(
    ext: &mut dyn Externalities,
    f: F,
) -> R
where
    F: FnOnce() -> R,

Set the given externalities while executing the given closure. To get access to the externalities while executing the given closure with_externalities grants access to them. The externalities are only set for the same thread this function was called from.