Struct sp_externalities::Extensions

pub struct Extensions { /* private fields */ }

Stores extensions that should be made available through the externalities.

Implementations

impl Extensions

pub fn new() -> Self

Create new instance of Self.

pub fn register<E: Extension>(&mut self, ext: E)

Register the given extension.

pub fn register_with_type_id( &mut self, type_id: TypeId, extension: Box<dyn Extension> ) -> Result<(), Error>

Register extension extension using the given type_id.

pub fn get_mut(&mut self, ext_type_id: TypeId) -> Option<&mut dyn Any>

Return a mutable reference to the requested extension.

pub fn deregister(&mut self, type_id: TypeId) -> bool

Deregister extension for the given type_id.

Returns true when the extension was registered.

pub fn iter_mut( &mut self ) -> impl Iterator<Item = (&TypeId, &mut Box<dyn Extension>)>

Returns a mutable iterator over all extensions.

pub fn merge(&mut self, other: Self)

Merge other into self.

If both contain the same extension, the extension instance of other will overwrite the instance found in self.