Crate sp_externalities

Substrate externalities abstraction

The externalities mainly provide access to storage and to registered extensions. Extensions are for example the keystore or the offchain externalities. These externalities are used to access the node from the runtime via the runtime interfaces.

This crate exposes the main Externalities trait.

Macros

decl_extension
Macro for declaring an extension that usable with Extensions.

Structs

Extensions
Stores extensions that should be made available through the externalities.

Enums

Error
Externalities error.

Traits

Extension
Marker trait for types that should be registered as Externalities extension.

ExtensionStore
Something that provides access to the Extensions store.

Externalities
The Substrate externalities.

ExternalitiesExt
Extension for the Externalities trait.

Functions

set_and_run_with_externalities
Set the given externalities while executing the given closure. To get access to the externalities while executing the given closure with_externalities grants access to them. The externalities are only set for the same thread this function was called from.

with_externalities
Execute the given closure with the currently set externalities.